pyproject.toml
[tool.poetry]
name = "easy-image-qr"
version = "0.1.0"
description = ""
authors = ["Tafeen <taafeenn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
qrcode = "^6.1"
pypng = "^0.0.20"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.8.3"
black = {version = "^21.5b2", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
